+++ httprunner/ext/locusts/__main__.py
@@ -1,0 +1,4 @@
+from httprunner.ext.locusts.cli import main
+
+if __name__ == "__main__":
+    main()
