--- httprunner/plugins/__init__.py
@@ -1,2 +1,0 @@
-# NOTICE:
-# This file should not be deleted, or ImportError will be raised in Python 2.7 when importing plugin
