--- httprunner/plugins/locusts/__main__.py
@@ -1,4 +1,0 @@
-from httprunner.plugins.locusts.cli import main
-
-if __name__ == "__main__":
-    main()
